make cla frontend use nested resources.
why? so we can log changes per case (not using nested resources would mean logging changes to an object e.g. personaldetails but there is no where to look at a log per object and all changes take place through the lens of a case.
this change means that you edit personaldetails through the case: /api/case/<case_ref>/personal_details/.

the benefit of this is that we also get rid of the associate_XXX_to_case calls from the frontend.