fixes missing angular sticky dependency
A previous PR removed sticky as the original angular-sticky repo was removed
from bower. I managed to find a suitable replacement for it however it
required Angular 1.4+ so as a result this commit updates Angular from 1.3
to 1.4. This meant that other dependencies also needed to be updated and
I had to refactor some of the code to get it to work without any errors.