allow providers can split cases
This addresses a bug discovered as a result of the angularjs version
being updated from 1.3 to 1.4. Providers reported to us that when they
tried to split a case the two matter type lists where not being populated
and were disabled.

At first it seemed as though the category drop down list was not triggering
the change event which in turn would populate the two matter type drop
down lists. The issue related to the scope that was being used, the
scope in the view and the scope watch being used in the controller were
different and thats why it was not working.

As part of this commit, I discovered that no tests were setup for this
feature or for any other features since mid-2015. The existing tests do
no work and I think this needs to be addresses as a bigger issue and
ticket. Therefore I have not added any tests.