fixes feedback checkboxes not working
User reported that they could not resolve feedback using the checkbox.

After investigating this issue, it was discovered the the click event
on the checkbox was being trigger twice. The reason for this is that
the checkbox had two two-way bindings between the models attribute.
By removing ng-model it resolved this issue. Also adding ng-checked allows
us to play back the checkboxes value to the user.