Only search for filter links
There are some instances when there are other links on the page that
similar text to the filters for eg. "EOD" this appears in the filters
and in the case list for Operator managers. The issue with not scoping
the search is that if the elements were moved around the test would
be searching and clicking on the wrong link and could return a false
positive.