Remove pushing images to abandoned Live-0 Cloud Platforms ECR repository
This project was pushing to old live-0 Cloud Platforms ECR repository which is no longer in use.
Cloud platforms have removed all non-production namespaces on live-0 ministryofjustice/cloud-platform-environments#1770

It seems the push to the ECR repository was added in anticipation of moving to kubernetes at the time.
However the Cloud platforms environments have changed since the original work.

When this project does move to cloud platforms managed environment, the relevant repository can be added.