Stop pages with a query string key "search" in the URL being sent to GA
This makes use of the $angularticsProvider, a service that gets injected into the application through adding angulartics and angulartics-google-analytics as dependencies.

The queryKeysBlacklist is a method that gets shipped after upgrading the angulartics library to v1.6. It targets any URLs that have a search parameter from being sent to google analytics.