Remove user data on refresh on the GA location property
The code finds the location property (URL) and if it contains a search query key, it splits the URL into two strings: the first string contains everything before the search query key and the second string contains everything after. It takes the first string and sets this as the new location.