Fix issue of screen not scrolling after clicking away from Finances tab
Code runs every time the Finances tab is clicked due to the 'onEnter' angular hook.

After the Finances tab is clicked, the Mutation Observer (responsible for detecting changes to the DOM/nodes) attaches an event listener to the tabs under Finances. After it disconnects and does not continue running in the background