Scroll scope form into view after clicking next for any of the questions
This code happens after every click on 'next' in the scope form. Mutation observer will run forever in the background, therefore it is 'disconnected' (removed) until the next click where it is re-added.

This keeps happening until eventually the form is completed and the next part is to fill in the 'Finances' section of the form. The Mutation Observer would have been disconnected by now, ensuring it is not running in the background forever.