Disable tab and prevent assign tab from being accessed
Disabling the tab just greys it out, it doesn't stop the tab from being accessed. So you would still be able to click on it and view the assign page.

The code in the js stops it from being accessed. It also writes a message on the banner at the top of the screen