Move angularUtils from bower to npm
AngularUtils is a project containing other modules. The specific module we care about here is the pagination module (called angular-utils-pagination in npm).

Since this is the only module we used, I installed the specific module instead of the whole project and installed the earliest version available which was 0.8.0 (we use 0.5.0 in bower).