Confirm whether we need to configure google analytics to remove error
There was an error being generated on the dash board page (sometimes on other pages e.g. case details) which was caused by the analytics.js file. It was trying to make a request to www.google.co.uk/ads/ga-audiences... which was getting blocked by CORS.

This was meant to fix that, but it was still appearing.

I cleared my browser cache and the issue seems to have been gone. Want to confirm clearing the cache fixes the issue